pyproject.toml
metadata-version = "2.3"
homepage = "https://github.com/volfpeter/mastodon-social-graph"
author = "Peter Volf"
author-email = "do.volfp@gmail.com"

[tool.black]
line-length = 118

[tool.mypy]
check_untyped_defs = true
show_error_codes = true

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages]
find = {}  # Scan the project directory with the default parameters

[project]
name = "mastodon-social-graph"
version = "0.2301.0"
description = "Mastodon social graph with an SQL backend, in-memory cache, and built-in, on-demand web scraper."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.8"
dependencies = ["graphscraper", "Mastodon.py"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Topic :: Database",
    "Topic :: Education",
    "Topic :: Internet",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Scientific/Engineering",
    "Topic :: Software Development :: Libraries",
    "Typing :: Typed"
]